refactor: move http/ out of legacy/ — refactored module no longer transitional
After #816 routed everything through Gateway and unified /message + /steer
into /dispatch, http/ is no longer "legacy waiting to be decomposed". Move
src/legacy/http/ → src/http/ to reflect first-class module status.

- mv src/legacy/http src/http
- Update relative imports inside (one less ../ level)
- Fix one outlier: status.ts still imports VERSION from src/legacy/version.ts
  (legacy/version.ts not yet moved)
- Update src/app.ts import path
- Refresh stale "src/plugins/http/..." path comments in file headers
- CLAUDE.md: drop http/ from legacy section, expand top-level entry
