+++ src/http/server.ts
@@ -1,11 +1,11 @@
-// src/plugins/http/server.ts — HTTP server for the Isotopes REST API
+// src/http/server.ts — HTTP server for the Isotopes REST API
 // Minimal server built on Node.js built-in http module (no Express).
 
 import http from "node:http";
 import path from "node:path";
-import { createLogger } from "../../logging/logger.js";
-import type { CronScheduler } from "../../automation/cron-job.js";
-import type { SessionStoreManager } from "../../agent/pi/session-store.js";
+import { createLogger } from "../logging/logger.js";
+import type { CronScheduler } from "../automation/cron-job.js";
+import type { SessionStoreManager } from "../agent/pi/session-store.js";
 import {
   applyCors,
   parseJsonBody,
@@ -16,7 +16,7 @@ import {
 } from "./middleware.js";
 import { matchRoute, type RouteDeps } from "./routes.js";
 import { serveStaticFile } from "./static.js";
-import { matchUIEntry, type UIEntry } from "../../extensions/ui/loader.js";
+import { matchUIEntry, type UIEntry } from "../extensions/ui/loader.js";
 
 import "./cron.js";
 import "./logs.js";
@@ -35,8 +35,8 @@ export interface ApiServerDeps {
   cronScheduler: CronScheduler;
   uiEntries?: UIEntry[];
   sessionStoreManager?: SessionStoreManager;
-  agentRuntime?: import("../../agent/runtime.js").AgentRuntime;
-  gateway?: import("../../gateway/index.js").Gateway;
+  agentRuntime?: import("../agent/runtime.js").AgentRuntime;
+  gateway?: import("../gateway/index.js").Gateway;
 }
 
 export class ApiServer {
